+++ commands/ops.py
@@ -1,4 +1,0 @@
-def deploy() -> dict:
-    """TODO: add description"""
-    # TODO: implement
-    return {"status": "ok"}